Update on project folder structure
